Refactor BackupRestore file structure
