+++ src/routes.js
@@ -6,7 +6,7 @@ import MangaInfo from "pages/MangaInfo";
 import Reader from "components/Reader";
 import Catalogue from "pages/Catalogue";
 import Extensions from "pages/Extensions";
-import BackupRestore from "pages/BackupRestore";
+import BackupRestore from "components/BackupRestore";
 import UrlPrefixContext from "components/UrlPrefixContext";
 import Settings, { SETTING_INDEX } from "components/Settings";
